Recommend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pandas==2.0.1
numpy==1.26.0
scikit-learn==1.2.2
pydantic==1.10.7
pytest==7.2.2
python-multipart==0.0.6
httpx==0.23.3
tqdm>=4.62.2
cf-step>=0.2.3
torch>=1.9.0
pgvector>=0.1.6
sqlalchemy==2.0.7
sentence-transformers==2.6.1
huggingface_hub>=0.22.0